+++ src/components/playlist/Setting/PlaylistSettingsDialog.tsx
@@ -4,7 +4,7 @@ import { Button, Dialog, Portal, Text, Switch } from 'react-native-paper';
 import { useTranslation } from 'react-i18next';
 
 import { useNoxSetting } from '@stores/useApp';
-import PortaledInput, { PortalInputRef } from '../dialogs/PortaledInput';
+import PortaledInput, { PortalInputRef } from '../../dialogs/PortaledInput';
 
 const styles = StyleSheet.create({
   switchContainer: {
